# openapi.json
{
  "components": {
    "schemas": {
...
      "AgentResumeRequest": {
        "properties": {
          "approved": {
            "default": true,
            "title": "Approved",
            "type": "boolean"
          },
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "AgentResumeRequest",
        "type": "object"
      },
      "AgentRunRequest": {
        "properties": {
          "graph": {
            "default": "support_agent",
            "title": "Graph",
            "type": "string"
          },
          "initial_state": {
            "additionalProperties": true,
            "title": "Initial State",
            "type": "object"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "recursion_limit": {
            "default": 25,
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Recursion Limit",
            "type": "integer"
          },
          "save_run": {
            "default": true,
            "title": "Save Run",
            "type": "boolean"
          }
        },
        "required": [
          "question"
        ],
        "title": "AgentRunRequest",
        "type": "object"
      },
      "AgentRunResponse": {
        "properties": {
          "diagnosis": {
            "items": {
              "type": "string"
            },
            "title": "Diagnosis",
            "type": "array"
          },
          "final_state": {
            "additionalProperties": true,
            "title": "Final State",
            "type": "object"
          },
          "graph": {
            "title": "Graph",
            "type": "string"
          },
          "halted_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Halted Reason"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/StepOut"
            },
            "title": "History",
            "type": "array"
          },
          "interrupted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupted At"
          },
          "looped": {
            "title": "Looped",
            "type": "boolean"
          },
          "node_visits": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Node Visits",
            "type": "object"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "steps": {
            "title": "Steps",
            "type": "integer"
          },
          "total_ms": {
            "title": "Total Ms",
            "type": "number"
          }
        },
        "required": [
          "graph",
          "status",
          "final_state",
          "history",
          "steps",
          "total_ms",
          "node_visits",
          "looped"
        ],
        "title": "AgentRunResponse",
        "type": "object"
      },
...
      "EvalRunRequest": {
        "properties": {
          "baseline_evaluation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Evaluation Id"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/EvalCaseIn"
            },
            "title": "Cases",
            "type": "array"
          },
          "config": {
            "$ref": "#/components/schemas/RAGConfigIn"
          },
          "name": {
            "default": "experiment",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "use_golden_set": {
            "default": true,
            "title": "Use Golden Set",
            "type": "boolean"
          },
          "use_llm_judge": {
            "default": false,
            "title": "Use Llm Judge",
            "type": "boolean"
          }
        },
        "title": "EvalRunRequest",
        "type": "object"
      },
...
      "GraphDiagramOut": {
        "properties": {
          "cycles": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Cycles",
            "type": "array"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdgeOut"
            },
            "title": "Edges",
            "type": "array"
          },
          "entry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry"
          },
          "info": {
            "additionalProperties": true,
            "title": "Info",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeOut"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "name",
          "entry",
          "nodes",
          "edges"
        ],
        "title": "GraphDiagramOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MentorResponseOut": {
        "properties": {
          "cost_usd": {
            "default": 0.0,
            "title": "Cost Usd",
            "type": "number"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "next_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Level"
          },
          "question_back": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Back"
          },
          "reveals_answer": {
            "title": "Reveals Answer",
            "type": "boolean"
          },
          "simulated": {
            "title": "Simulated",
            "type": "boolean"
          },
          "tokens": {
            "default": 0,
            "title": "Tokens",
            "type": "integer"
          },
          "trace": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Trace",
            "type": "array"
          }
        },
        "required": [
          "level",
          "message",
          "reveals_answer",
          "simulated"
        ],
        "title": "MentorResponseOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\n**AI Forge Labs** \u2014 a Python & AI engineering RPG.\n\nLearn it. Build it. Break it. Debug it. Explain it. Master it.\n\nThe API behind an engineering simulator that trains Python, backend, data, ML,\ndeep learning, LLM, RAG and agent skills through missions, debugging incidents,\nproduction war rooms and adaptive interviews \u2014 with a spaced-repetition engine\nthat brings back what you are about to forget.\n",
    "title": "AI Forge Labs",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/labs/agents/resume": {
      "post": {
        "description": "Approve or reject a run paused at a human-in-the-loop checkpoint.",
        "operationId": "resume_agent_api_v1_labs_agents_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resume Agent Api V1 Labs Agents Resume Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resume Agent",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/agents/run": {
      "post": {
        "description": "Execute a graph and return the state diff at every step.",
        "operationId": "run_agent_api_v1_labs_agents_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Agent",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/agents/{slug}": {
      "get": {
        "description": "The topology, **before** running anything.\n\nInspecting a graph without executing it is the difference between debugging\na diagram and debugging a stack trace. Cycles are reported \u2014 a cycle is not\na bug, but every one needs a termination argument.",
        "operationId": "graph_diagram_api_v1_labs_agents__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphDiagramOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Graph Diagram",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/evals/run": {
      "post": {
        "description": "Run a configuration against a golden set and decide ship / hold / rollback.\n\nRegression against a baseline outranks absolute thresholds: a system that\ngot worse is a rollback even when it still clears the gate, because the next\nchange will make it worse again.",
        "operationId": "run_evaluation_api_v1_labs_evals_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Evaluation",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/mentor/ask": {
      "post": {
        "description": "Escalating help \u2014 never the answer before it is earned (spec \u00a758).\n\nThe response includes the mentor's own graph trace, so the tool that is\nhelping you can be inspected in the same Agent Inspector it is built on.",
        "operationId": "ask_mentor_api_v1_labs_mentor_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MentorAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorResponseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ask Mentor",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/mentor/generate": {
      "post": {
        "description": "Generate practice questions from a concept \u2014 and validate them.\n\nGenerated content is held to the *same* bar as hand-authored content,\nincluding the check that the generated ideal answer scores at least 0.6\ncorrectness against its own generated rubric. Rejected questions are\nreturned with their reasons rather than hidden, because seeing what a model\ngets wrong about question design is itself instructive.",
        "operationId": "generate_questions_api_v1_labs_mentor_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateQuestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Questions",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/mentor/socratic": {
      "post": {
        "description": "Ask the follow-up that reveals whether the player understood or pattern-matched.",
        "operationId": "socratic_api_v1_labs_mentor_socratic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocraticRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocraticResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Socratic",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/rag/chunk-preview": {
      "post": {
        "description": "Preview chunking without embedding anything.\n\nFast enough to drive a live slider, which is the point: seeing chunk\nboundaries move as you drag is how the trade-off becomes intuitive.",
        "operationId": "chunk_preview_api_v1_labs_rag_chunk_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChunkPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chunk Preview",
        "tags": [
          "labs"
        ]
      }
    },
...
    "/api/v1/labs/rag/experiments": {
      "get": {
        "description": "Past runs, with the delta between the two most recent and a ship verdict.",
        "operationId": "rag_experiments_api_v1_labs_rag_experiments_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGCompareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rag Experiments",
        "tags": [
          "labs"
        ]
      }
    },
...
    "/api/v1/labs/rag/query": {
      "post": {
        "description": "Run the full pipeline and return every stage, every score, and a diagnosis.",
        "operationId": "rag_query_api_v1_labs_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGQueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rag Query",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/status": {
      "get": {
        "description": "Which labs are fully live versus running on the offline stubs.",
        "operationId": "lab_status_api_v1_labs_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Lab Status Api V1 Labs Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Lab Status",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/transformer/attention": {
      "post": {
        "description": "Compute one multi-head self-attention block and return every intermediate.\n\nEvery number is really computed \u2014 real embeddings, real positional encoding,\nreal Q\u00b7K\u1d40/\u221ad, real softmax. The projections are deterministic pseudo-random\nrather than trained, and the response says so: the mechanism is genuine, the\nlearned behaviour is not.",
        "operationId": "attention_api_v1_labs_transformer_attention_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttentionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttentionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attention",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/transformer/sampling": {
      "post": {
        "description": "Show what temperature, top-k and top-p do to the same distribution.",
        "operationId": "sampling_api_v1_labs_transformer_sampling_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamplingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SamplingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sampling",
        "tags": [
          "labs"
        ]
      }
    },
    "/api/v1/labs/transformer/tokenize": {
      "post": {
        "description": "Tokenize text and price it.\n\nCost is reported per million requests as well as per request, because\nper-request pennies are exactly why token bloat goes unnoticed until the\ninvoice arrives.",
        "operationId": "tokenize_api_v1_labs_transformer_tokenize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tokenize",
        "tags": [
          "labs"
        ]
      }
    },
...
  },
...
}